+++ mek_data_class_generator/CHANGELOG.md
@@ -1,2 +1,6 @@
+## 0.0.3
+- Expose `update` method in `*Changes` class to update a class with function
+- Expose `replace` method to update `*Changes` class with all DataClass properties
+
 ## 0.0.1
 - Initial version.